refactor: Header 클래스를 따로 분리
 - 각각의 class를 작은 단위로 유지하기 위해
