+++ README.md
@@ -7,6 +7,7 @@
 - [x] 역 관리 레이아웃 잡기
 - [x] 역 추가 및 삭제 하기(예외 상황 처리)
 - [x] 노선 관리 레이아웃 잡기
+- [ ] header 모듈 분리하기
 - [ ] 상행 하행 종점에 역 리스트 넣기
 - [ ] 노선 추가 및 삭제 하기
 - [ ] 구간 관리 레이아웃 잡기
